+++ lib/src/models/activity_record_model.g.dart
@@ -9,11 +9,11 @@ part of 'activity_record_model.dart';
 ActivityRecordModel _$ActivityRecordModelFromJson(Map<String, dynamic> json) =>
     ActivityRecordModel(
       digital: json['digital'] as String,
-      hours: json['hours'] as int,
-      minutes: json['minutes'] as int,
+      hours: (json['hours'] as num).toInt(),
+      minutes: (json['minutes'] as num).toInt(),
       name: json['name'] as String,
       percent: (json['percent'] as num).toDouble(),
-      seconds: json['seconds'] as int,
+      seconds: (json['seconds'] as num).toInt(),
       text: json['text'] as String,
       totalSeconds: (json['total_seconds'] as num).toDouble(),
     );
@@ -31,8 +31,8 @@ Map<String, dynamic> _$ActivityRecordModelToJson(
       'total_seconds': instance.totalSeconds,
     };
 
-Map<String, dynamic> _$$_ActivityRecordModelToJson(
-        _$_ActivityRecordModel instance) =>
+Map<String, dynamic> _$$ActivityRecordModelImplToJson(
+        _$ActivityRecordModelImpl instance) =>
     <String, dynamic>{
       'digital': instance.digital,
       'hours': instance.hours,