+++ lib/src/models/grand_total_model.g.dart
@@ -10,8 +10,8 @@ GrandTotalModel _$GrandTotalModelFromJson(Map<String, dynamic> json) =>
     GrandTotalModel(
       decimal: json['decimal'] as String,
       digital: json['digital'] as String,
-      hours: json['hours'] as int,
-      minutes: json['minutes'] as int,
+      hours: (json['hours'] as num).toInt(),
+      minutes: (json['minutes'] as num).toInt(),
       text: json['text'] as String,
       totalSeconds: (json['total_seconds'] as num).toDouble(),
     );
@@ -26,7 +26,8 @@ Map<String, dynamic> _$GrandTotalModelToJson(GrandTotalModel instance) =>
       'total_seconds': instance.totalSeconds,
     };
 
-Map<String, dynamic> _$$_GrandTotalModelToJson(_$_GrandTotalModel instance) =>
+Map<String, dynamic> _$$GrandTotalModelImplToJson(
+        _$GrandTotalModelImpl instance) =>
     <String, dynamic>{
       'decimal': instance.decimal,
       'digital': instance.digital,